Relocate `@CreateIfMissing` to public pkg & Add javadoc
