+++ core/src/main/java/com/linecorp/armeria/server/annotation/ServerSentEventResponseConverterFunction.java
@@ -32,7 +32,6 @@ import com.linecorp.armeria.common.MediaType;
 import com.linecorp.armeria.common.ResponseHeaders;
 import com.linecorp.armeria.common.annotation.Nullable;
 import com.linecorp.armeria.common.sse.ServerSentEvent;
-import com.linecorp.armeria.internal.common.CreateIfMissing;
 import com.linecorp.armeria.server.ServiceRequestContext;
 
 /**
