+++ core/src/main/java/com/linecorp/armeria/server/annotation/decorator/RateLimitingDecoratorFactoryFunction.java
@@ -18,7 +18,7 @@ package com.linecorp.armeria.server.annotation.decorator;
 import java.util.function.Function;
 
 import com.linecorp.armeria.common.HttpRequest;
-import com.linecorp.armeria.internal.common.CreateIfMissing;
+import com.linecorp.armeria.server.annotation.CreateIfMissing;
 import com.linecorp.armeria.internal.server.annotation.DefaultValues;
 import com.linecorp.armeria.server.HttpService;
 import com.linecorp.armeria.server.annotation.DecoratorFactoryFunction;
