+++ core/src/main/java/com/linecorp/armeria/server/cors/CorsDecoratorsFactoryFunction.java
@@ -21,7 +21,7 @@ import static java.util.Objects.requireNonNull;
 import java.util.Arrays;
 import java.util.function.Function;
 
-import com.linecorp.armeria.internal.common.CreateIfMissing;
+import com.linecorp.armeria.server.annotation.CreateIfMissing;
 import com.linecorp.armeria.server.HttpService;
 import com.linecorp.armeria.server.annotation.DecoratorFactoryFunction;
 import com.linecorp.armeria.server.annotation.decorator.CorsDecorator;
